backend/requirements.txt
# Web Framework
FastAPI==0.103.2
uvicorn[standard]==0.23.2
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-RESTX==1.1.0

# Database
PyMongo==4.5.0
Motor==3.3.2

# Data Processing & Validation
pydantic==2.4.2
pydantic[email]==2.4.2
pydantic-settings==2.0.3
jsonschema==4.19.1
marshmallow==3.20.1
json-source-map>=0.6.0
networkx>=3.0

# Utilities
python-dotenv==1.0.0
python-magic==0.4.27
requests==2.31.0
click==8.1.7

# Authentication & Security
PyJWT==2.8.0
bcrypt==5.0.0
python-multipart==0.0.6

# Testing frameworks
pytest>=7.4.2
pytest-flask>=1.2.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
selenium>=4.15.0
webdriver-manager>=4.0.1
responses>=0.23.3

# Code Quality
ruff>=0.1.0
pre-commit>=3.4.0

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0

# Development tools
tomli-w>=1.0.0
